+++ src/main/java/com/thiendz/tool/fplautocms/dto/GetQuizQuestionDto.java
@@ -1,0 +1,9 @@
+package com.thiendz.tool.fplautocms.dto;
+
+import lombok.Data;
+import lombok.ToString;
+
+@Data
+@ToString
+public class GetQuizQuestionDto extends ResponseDto<DataB64QuizDto> {
+}