+++ src/main/java/com/thiendz/tool/fplautocms/dto/callback/CallbackSolution.java
@@ -1,8 +1,0 @@
-package com.thiendz.tool.fplautocms.dto.callback;
-
-import com.thiendz.tool.fplautocms.models.Quiz;
-
-@FunctionalInterface
-public interface CallbackSolution {
-    void call(double scorePresent, int status, Quiz quiz);
-}