+++ src/main/java/com/thiendz/tool/fplautocms/models/QuizQuestion.java
@@ -1,6 +1,7 @@
 package com.thiendz.tool.fplautocms.models;
 
 
+import com.thiendz.tool.fplautocms.utils.NumberUtils;
 import com.thiendz.tool.fplautocms.utils.enums.QuizQuestionType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,20 +14,26 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class QuizQuestion {
-    private String name;
+public class QuizQuestion implements Comparable<QuizQuestion> {
     private QuizQuestionType type;
+    private String name;
     private String question;
-    private String key; //key request
-    private List<String> listValue;// value request
-    //số lượng câu hỏi trên một question (dành cho type text)
-    private int amountInput;
-    //có chọn nhiều đáp ans ko
+    private String key;
+    private List<String> value;
+    private List<String> listValue;
+    private int input;
     private boolean multiChoice;
-    //giá trị request đi
-    private String selectValue;
-    //số lần thử
-    private int testCount;
-    //hoàn thành
+    private int test;
     private boolean correct;
+
+    @Override
+    public int compareTo(QuizQuestion thatQuizQuestion) {
+        Integer thisQuizNumber = NumberUtils.getInt(this.getName());
+        Integer thatQuizNumber = NumberUtils.getInt(thatQuizQuestion.getName());
+        if (thisQuizNumber == null)
+            return -1;
+        if (thatQuizNumber == null)
+            return 1;
+        return thisQuizNumber - thatQuizNumber;
+    }
 }