+++ src/main/java/com/thiendz/tool/fplautocms/services/SolutionService.java
@@ -1,17 +1,17 @@
 package com.thiendz.tool.fplautocms.services;
 
-import com.thiendz.tool.fplautocms.dto.callback.CallbackSolution;
 import com.thiendz.tool.fplautocms.models.Course;
 import com.thiendz.tool.fplautocms.models.Quiz;
 import com.thiendz.tool.fplautocms.models.QuizQuestion;
 import com.thiendz.tool.fplautocms.models.User;
 import com.thiendz.tool.fplautocms.dto.SolutionResponseDto;
 import com.thiendz.tool.fplautocms.utils.*;
-import com.thiendz.tool.fplautocms.utils.consts.Messages;
+import com.thiendz.tool.fplautocms.utils.consts.UserAgent;
 import com.thiendz.tool.fplautocms.utils.enums.QuizQuestionType;
 import com.thiendz.tool.fplautocms.utils.excepts.CmsException;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.http.client.HttpClient;
+import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.fluent.Executor;
 import org.apache.http.client.fluent.Request;
 import org.apache.http.entity.ContentType;
@@ -21,7 +21,9 @@ import org.jsoup.nodes.Document;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 public class SolutionService implements Runnable {
@@ -32,23 +34,33 @@ public class SolutionService implements Runnable {
     private final User user;
     private final Course course;
     private final Quiz quiz;
-
-    private CallbackSolution callbackSolution;
+    private String paramPost;
+    private boolean resetScoreQuiz;
 
     private int status;
-    private double scorePresent;
+    private Double scorePresent;
 
     public SolutionService(User user, Course course, Quiz quiz) {
         this.user = user;
         this.course = course;
         this.quiz = quiz;
         this.scorePresent = quiz.getScore();
-        this.callbackSolution = ((scorePresent, status, quiz1) -> {
-        });
+    }
+
+    public void setResetScoreQuiz(boolean resetScoreQuiz) {
+        this.resetScoreQuiz = resetScoreQuiz;
+    }
+
+    public void setParamPost(String paramPost) {
+        this.paramPost = paramPost;
     }
 
     public Quiz getQuiz() {
         return quiz;
+    }
+
+    public String getParamPost() {
+        return paramPost;
     }
 
     public double getScorePresent() {
@@ -59,10 +71,6 @@ public class SolutionService implements Runnable {
         return status;
     }
 
-    public void setCallbackSolution(CallbackSolution callbackSolution) {
-        this.callbackSolution = callbackSolution;
-    }
-
     @Override
     public void run() {
         try {
@@ -71,7 +79,7 @@ public class SolutionService implements Runnable {
             log.info("Quiz [{}] - score: [{}] solution finish.", quiz.getName(), scorePresent);
         } catch (CmsException | IOException | NullPointerException ex) {
             setStatus(-1);
-            log.error("Solution error: {}", ex.toString());
+            log.error("Solution error!", ex);
         }
     }
 
@@ -89,8 +97,12 @@ public class SolutionService implements Runnable {
         );
         do {
             if (DateUtils.getCurrentMilis() - timeTick > TIME_SLEEP_SOLUTION) {
-                String bodyParam = buildParam();
-                final HttpClient client = HttpClientBuilder.create()
+                String bodyParam = paramPost == null ? buildParam() : paramPost;
+                RequestConfig requestConfig = RequestConfig.custom()
+                        .setConnectTimeout(60000)
+                        .build();
+                HttpClient client = HttpClientBuilder.create()
+                        .setDefaultRequestConfig(requestConfig)
                         .disableRedirectHandling()
                         .build();
                 Executor executor = Executor.newInstance(client);
@@ -99,14 +111,12 @@ public class SolutionService implements Runnable {
                         .setHeader("Referer", quiz.getUrl())
                         .setHeader("Accept", "application/json, text/javascript, */*; q=0.01")
                         .setHeader("Cookie", user.getCookie())
-                        .setHeader("User-Agent", Messages.APP_NAME + " v" + Messages.APP_VER)
-                        .setHeader("AutoBy", Messages.APP_AUTHOR)
-                        .setHeader("Contact", Messages.APP_CONTACT)
+                        .setHeader("User-Agent", UserAgent.USER_AGENT_DEFAULT)
                         .bodyString(bodyParam, ContentType.create("application/x-www-form-urlencoded", StandardCharsets.UTF_8));
                 String bodyResponseSolution = executor.execute(request).returnContent().asString();
                 SolutionResponseDto solutionResponseDto = MapperUtils.objectMapper.readValue(bodyResponseSolution, SolutionResponseDto.class);
-                scorePresent = solutionResponseDto.getCurrent_score();
-                updateStatusQuizQuestion(solutionResponseDto.getContents());
+                setScorePresent(solutionResponseDto.getCurrent_score());
+                updateStatusQuizQuestion(solutionResponseDto);
                 log.info("Quiz: {}", quiz.getName());
                 log.info("Request POST: {}", url);
                 log.info("Request Send: {}", bodyParam);
@@ -122,94 +132,80 @@ public class SolutionService implements Runnable {
 
     private String buildParam() throws CmsException {
         List<QuizQuestion> quizQuestionList = quiz.getQuizQuestions();
-        StringBuilder sb = new StringBuilder();
+        List<String> paramList = new ArrayList<>();
         for (int i = 0; i < quizQuestionList.size(); i++) {
             QuizQuestion quizQuestion = quizQuestionList.get(i);
-            if (quizQuestion.getListValue() == null) {
+            if (quizQuestion.getListValue() == null)
                 continue;
-            }
-            String ans = setValue(quizQuestionList.get(i));
-            sb.append(ans).append("&");
+            List<String> selectNextValueList = selectNextValue(quizQuestion);
+            List<String> selectNextValueEncryptList = selectNextValueList
+                    .stream().map(StringUtils::URLEncoder)
+                    .collect(Collectors.toList());
+            String param;
+            String keyEncrypt = StringUtils.URLEncoder(quizQuestion.getKey());
+            if (quizQuestion.getType() == QuizQuestionType.TEXT)
+                param = keyEncrypt + "=" + String.join("%2C", selectNextValueEncryptList);
+            else
+                param = selectNextValueEncryptList
+                        .stream().map(s -> keyEncrypt + "=" + s)
+                        .collect(Collectors.joining("&"));
             if (!quizQuestion.isCorrect()) {
-                quiz.getQuizQuestions().get(i).setTestCount(quiz.getQuizQuestions().get(i).getTestCount() + 1);
+                quiz.getQuizQuestions().get(i).setTest(quizQuestion.getTest() + 1);
+                quiz.getQuizQuestions().get(i).setValue(selectNextValueList);
             }
-            quiz.getQuizQuestions().get(i).setSelectValue(ans);
+            paramList.add(param);
         }
-        return makeUpValue(sb.toString());
+        return String.join("&", paramList);
     }
 
-    private String setValue(QuizQuestion quizQuestion) throws CmsException {
-        if (quizQuestion.isCorrect()) {
-            return quizQuestion.getSelectValue();
-        }
-        String keyEncrypt = StringUtils.URLEncoder(quizQuestion.getKey());
+    private List<String> selectNextValue(QuizQuestion quizQuestion) throws CmsException {
+        if (quizQuestion.isCorrect())
+            return quizQuestion.getValue();
+        int next = quizQuestion.getTest();
+        List<String> nextValueList = new ArrayList<>();
         if (quizQuestion.isMultiChoice()) {
-            StringBuilder value = new StringBuilder();
-            List<List<Integer>> alInt;
-            if (quizQuestion.getType() == QuizQuestionType.TEXT) {
-                alInt = new Permutation(quizQuestion.getAmountInput(), quizQuestion.getListValue().size()).getResult();
-            } else {
-                alInt = new Combination(2, quizQuestion.getListValue().size(), true).getResult();
-            }
-            int index = quizQuestion.getTestCount();
-            if (index >= alInt.size()) {
-                throw new CmsException("ArrayIndexOutOfBound alInt!");
-            }
-            if (quizQuestion.getType() == QuizQuestionType.TEXT) {
-                value.append(keyEncrypt).append("=");
-            }
-            alInt.get(index).forEach((i) -> {
-                String valueEncrypt = StringUtils.URLEncoder(quizQuestion.getListValue().get(i));
-                if (quizQuestion.getType() == QuizQuestionType.TEXT) {
-                    value.append(valueEncrypt).append("%2C");
-                } else {
-                    value.append(keyEncrypt).append("=").append(valueEncrypt).append("&");
-                }
-            });
-            return makeUpValue(value.toString());
+            List<List<Integer>> selectList;
+            if (quizQuestion.getType() == QuizQuestionType.TEXT)
+                selectList = new Permutation(quizQuestion.getInput(), quizQuestion.getListValue().size()).getResult();
+            else
+                selectList = new Combination(2, quizQuestion.getListValue().size(), true).getResult();
+            if (next > selectList.size() - 1)
+                throw new CmsException("[M]Câu trả lời nằm ngoài danh sách đáp án có sẵn.");
+            selectList.get(next).forEach(select -> nextValueList.add(quizQuestion.getListValue().get(select)));
         } else {
-            List<String> choiceList = quizQuestion.getListValue();
-            String valueEncrypt = StringUtils.URLEncoder(choiceList.get(quizQuestion.getTestCount()));
-            String res = keyEncrypt + "=" + valueEncrypt + "&";
-            return makeUpValue(res);
+            if (next > quizQuestion.getListValue().size() - 1)
+                throw new CmsException("[S]Câu trả lời nằm ngoài danh sách đáp án có sẵn.");
+            nextValueList.add(quizQuestion.getListValue().get(next));
         }
+        return nextValueList;
     }
 
-    private void updateStatusQuizQuestion(String body) throws CmsException {
-        Document document = Jsoup.parse(body);
+
+    private void updateStatusQuizQuestion(SolutionResponseDto solutionResponseDto) throws CmsException {
+        if (solutionResponseDto.getContents() == null)
+            throw new CmsException(solutionResponseDto.getSuccess());
+        Document document = Jsoup.parse(solutionResponseDto.getContents());
         List<QuizQuestion> quizResults = QuizDetailService.buildQuizQuestions(document, true);
-        if (!compareKeyQuizQuestion(quiz.getQuizQuestions(), quizResults)) {
-            throw new CmsException("QuizResult != QuizStandard!");
-        }
-        for (int i = 0; i < quizResults.size(); i++) {
+        if (!compareKeyQuizQuestion(quiz.getQuizQuestions(), quizResults))
+            throw new CmsException("Dữ liệu trả về không khớp dữ liệu tại máy.");
+        for (int i = 0; i < quizResults.size(); i++)
             quiz.getQuizQuestions().get(i).setCorrect(quizResults.get(i).isCorrect());
-        }
     }
 
     private boolean compareKeyQuizQuestion(List<QuizQuestion> quizQuestionList1, List<QuizQuestion> quizQuestionList2) {
         if (quizQuestionList1.size() != quizQuestionList2.size())
             return false;
-        for (int i = 0; i < quizQuestionList1.size(); i++) {
+        for (int i = 0; i < quizQuestionList1.size(); i++)
             if (!quizQuestionList1.get(i).getKey().equals(quizQuestionList2.get(i).getKey()))
                 return false;
-        }
         return true;
-    }
-
-    private String makeUpValue(String value) {
-        int len = value.length();
-        if (value.endsWith("&")) {
-            return value.substring(0, len - 1);
-        }
-        if (value.endsWith("%2C")) {
-            return value.substring(0, len - 3);
-        }
-        return value;
     }
 
     private boolean isQuizFinished() {
         List<QuizQuestion> quizQuestionList = quiz.getQuizQuestions();
-        boolean maxScore = (quiz.getScore() == quiz.getScorePossible() && quiz.getScore() != 0);
+        boolean maxScore = false;
+        if (!resetScoreQuiz)
+            maxScore = quiz.getScore() == quiz.getScorePossible() && quiz.getScore() != 0;
         boolean quizFinish = quizQuestionList.stream()
                 .filter(quizQuestion -> quizQuestion.isCorrect() || quizQuestion.getListValue() == null)
                 .count() == quizQuestionList.size();
@@ -217,14 +213,20 @@ public class SolutionService implements Runnable {
     }
 
     private void resetQuizQuestion() {
-        for (QuizQuestion quizQuestion : quiz.getQuizQuestions()) {
-            quizQuestion.setCorrect(false);
+        for (int i = 0; i < quiz.getQuizQuestions().size(); i++) {
+            quiz.getQuizQuestions().get(i).setCorrect(false);
+            quiz.getQuizQuestions().get(i).setTest(0);
+            quiz.getQuizQuestions().get(i).setValue(null);
         }
     }
 
     private void setStatus(int status) {
         this.status = status;
-        callbackSolution.call(scorePresent, status, quiz);
+    }
+
+    private void setScorePresent(Double scorePresent) {
+        this.scorePresent = scorePresent;
+        quiz.setScore(scorePresent);
     }
 }