+++ src/main/java/com/thiendz/tool/fplautocms/utils/consts/Environments.java
@@ -4,4 +4,6 @@ package com.thiendz.tool.fplautocms.utils.consts;
 public class Environments {
     public static boolean DISABLE_ANALYSIS = false;
     public static boolean SAVE_LOG = false;
+    public static String SERVER_ADDRESS = "https://poly.11x7.xyz";
+    public static boolean ENABLE_RESET_QUIZ = false;
 }