+++ change-log.md
@@ -1,4 +1,13 @@
 # Change Log
+### v4.0.1
+Fake request trách việc phát bị phát hiện bởi nhà trường.  
+Fix write log use log4j.  
+Add log error in logs/error.log.  
+Fix connectionTimeout (60s).  
+Add env `-sa` or `--serve-add`.  
+Add env `-erq` or `--enable-reset-quiz`.  
+Auto 10 quiz 10 point 10 sec.  
+
 ### v4.0.0
 Chuyển sang sử dụng maven project.  
 Nâng cấp toàn bộ core hệ thống.